[filter] Implemented filter on single machine.
Revised the getRootOp() interface so that NodeServer parameter is not
passed in. Instead, the NodeServer parameter is passed into the
constructor of QScan, which is the only operation that needs to be aware
of the NodeServer.

This consequently requires the parse() method of QueryParser() to receive
a NodeServer parameter. Passing in null for this is allowed and does not
throw an exception. However, getRootOp() will throw a NullPointerException
since there is no reasonable query plan when there is no NodeServer.