[join] Implemented Join. Bushy joins seem to work as well.
The column indexes seem to be a bit awkward though. This is because the
user needs to be aware of what the index is for the column of the table
they are operating on. Especially for bushy joins, the user needs to know
the index after a bushy join, which can make the index really high.
Regardless, it seems a join can be computed, and this is sufficient for
our purposes.